Move svg jsx components to icons directory
