+++ taui/src/icons/map-marker-icon.js
