+++ taui/src/icons/polygon-icon.js
